requirements.txt
tqdm
pillow
omegaconf
matplotlib
tensorboard
lpips==0.1.4
torch~=1.13.1
torchvision~=0.14.1
git+https://github.com/toshas/torch-fidelity.git@master#egg=torch-fidelity
